+++ common/src/main/java/com/verisign/vscc/hdfs/trumpet/utils/TrumpetHelper.java
@@ -2,16 +2,9 @@ package com.verisign.vscc.hdfs.trumpet.utils;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.verisign.vscc.hdfs.trumpet.kafka.KafkaUtils;
-import com.verisign.vscc.hdfs.trumpet.kafka.SimpleConsumerHelper;
 import kafka.message.Message;
-import org.apache.curator.framework.CuratorFramework;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
 
 
@@ -27,4 +20,5 @@ public class TrumpetHelper {
     public static Map<String, Object> toMap(Message message) throws IOException {
         return toMap(KafkaUtils.toByteArray(message));
     }
+
 }