+++ common/src/test/java/com/verisign/vscc/hdfs/trumpet/kafka/SetupSimpleKafkaCluster.java
@@ -3,16 +3,22 @@ package com.verisign.vscc.hdfs.trumpet.kafka;
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.collect.Iterables;
+import kafka.cluster.EndPoint;
+import kafka.cluster.EndPoint$;
 import kafka.server.KafkaConfig;
 import kafka.server.KafkaServer;
 import kafka.utils.*;
 import kafka.zk.EmbeddedZookeeper;
 import org.I0Itec.zkclient.ZkClient;
+import org.I0Itec.zkclient.ZkConnection;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
 import org.apache.curator.retry.ExponentialBackoffRetry;
+import org.apache.kafka.common.protocol.SecurityProtocol;
+import org.apache.kafka.common.security.JaasUtils;
 import org.junit.After;
 import org.junit.Before;
+import scala.Option$;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
@@ -30,7 +36,9 @@ public abstract class SetupSimpleKafkaCluster {
 
     protected String zkConnect;
     protected EmbeddedZookeeper zkServer;
+    protected ZkConnection zkConnection;
     protected ZkClient zkClient;
+    protected ZkUtils zkUtils;
     protected KafkaServer kafkaServer;
     protected String kafkaServersString = "";
     protected List<KafkaServer> servers = new ArrayList<>();
@@ -41,19 +49,24 @@ public abstract class SetupSimpleKafkaCluster {
     public void setup() throws Exception {
 
         // setup Zookeeper
-        zkConnect = TestZKUtils.zookeeperConnect();
-        zkServer = new EmbeddedZookeeper(zkConnect);
-        zkClient = new ZkClient(zkServer.connectString(), zkConnectionTimeout, zkSessionTimeout, ZKStringSerializer$.MODULE$);
+        zkServer = new EmbeddedZookeeper();
+        zkConnect = "127.0.0.1:" + zkServer.zookeeper().getClientPort();
+
+        zkConnection = new ZkConnection(zkConnect,zkSessionTimeout);
+        zkClient = new ZkClient(zkConnection, zkConnectionTimeout, ZKStringSerializer$.MODULE$);
+        zkUtils = new ZkUtils(zkClient, zkConnection, JaasUtils.isZkSecurityEnabled());
 
         // setup Broker
-        int port = TestUtils.choosePort();
-        Properties props = TestUtils.createBrokerConfig(brokerId, port, true);
+        int port = TestUtils.RandomPort();
+        Properties props = TestUtils.createBrokerConfig(brokerId, zkConnect, false, false,
+                port, Option$.MODULE$.empty(), Option$.MODULE$.empty(), Option$.MODULE$.empty(), true, false, -1,
+                false, -1, false, -1, Option$.MODULE$.empty());
 
         KafkaConfig config = new KafkaConfig(props);
         Time mock = new MockTime();
         kafkaServer = TestUtils.createServer(config, mock);
         servers.add(kafkaServer);
-        kafkaServersString += kafkaServer.socketServer().host() + ":" + kafkaServer.socketServer().port();
+        kafkaServersString = kafkaServersToListOfString(servers);
 
         curatorFramework = CuratorFrameworkFactory.builder().connectString(zkConnect)
                 .retryPolicy(new ExponentialBackoffRetry(1000, 3))
@@ -88,7 +101,7 @@ public abstract class SetupSimpleKafkaCluster {
             @Nullable
             @Override
             public String apply(KafkaServer input) {
-                return input.socketServer().host() + ":" + input.socketServer().port();
+                return "127.0.0.1:" + input.boundPort(SecurityProtocol.PLAINTEXT);
             }
         }));
     }