+++ server/src/main/bin/peek-events.sh
@@ -64,13 +64,18 @@ fi
 #JAVA_OPTS="-XX:+HeapDumpOnOutOfMemoryError -verbose:gc"
 
 # Start up size for memory allocation pool for java VM.
-MS=512m
+MS=256m
 
 # Max size of memory allocation pool for java VM.
 MX=2048m
 
 JAVA_OPTS="$JAVA_OPTS -Xms$MS -Xmx$MX -XX:+UseParNewGC -XX:+UseConcMarkSweepGC -XX:-CMSConcurrentMTEnabled -XX:CMSInitiatingOccupancyFraction=70 -XX:+CMSParallelRemarkEnabled -XX:+DoEscapeAnalysis"
-JAVA_OPTS="$JAVA_OPTS -Dlogback.configurationFile=logback-to-console.xml"
+JAVA_OPTS="$JAVA_OPTS -Dlog4j.configuration=log4j-to-console.properties"
+
+if [ -f $HOMEDIR/config/kafka_client_jaas.conf ]
+then
+    JAVA_OPTS="${JAVA_OPTS} -Djava.security.auth.login.config=$HOMEDIR/config/kafka_client_jaas.conf"
+fi
 
 export HADOOP_USER_CLASSPATH_FIRST=true