+++ server/src/main/java/com/verisign/vscc/hdfs/trumpet/server/TrumpetLeader.java
@@ -1,13 +1,10 @@
 package com.verisign.vscc.hdfs.trumpet.server;
 
-import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Preconditions;
 import com.verisign.vscc.hdfs.trumpet.dto.EventAndTxId;
 import com.verisign.vscc.hdfs.trumpet.kafka.SimpleConsumerHelper;
 import com.verisign.vscc.hdfs.trumpet.server.editlog.EditLogDir;
-import com.verisign.vscc.hdfs.trumpet.server.editlog.WatchDog;
 import com.verisign.vscc.hdfs.trumpet.server.metrics.Metrics;
 import com.verisign.vscc.hdfs.trumpet.server.rx.EditLogObservable;
 import com.verisign.vscc.hdfs.trumpet.server.rx.ProducerSubscriber;
@@ -16,11 +13,8 @@ import kafka.message.Message;
 import org.apache.avro.util.ByteBufferInputStream;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.recipes.leader.CancelLeadershipException;
-import org.apache.curator.framework.recipes.leader.LeaderSelector;
 import org.apache.curator.framework.recipes.leader.LeaderSelectorListener;
 import org.apache.curator.framework.recipes.leader.LeaderSelectorListenerAdapter;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.hdfs.DFSConfigKeys;
 import org.apache.hadoop.hdfs.DistributedFileSystem;
 import org.codehaus.jackson.JsonNode;
@@ -29,14 +23,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import rx.Observable;
 import rx.Subscription;
-import rx.functions.Func1;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
 import java.util.Collections;
-import java.util.Map;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
@@ -54,6 +46,7 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
     private final String topic;
     private final EditLogDir editLogDir;
     private final long baseThrottleTimeMs;
+    private final int kafkaRequiredAcks;
 
     private final DistributedFileSystem dfs;
 
@@ -64,15 +57,16 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
 
     public TrumpetLeader(CuratorFramework curatorFramework, DistributedFileSystem dfs, String topic, EditLogDir editLogDir)
             throws IOException {
-        this(curatorFramework, dfs, topic, editLogDir, DEFAULT_BASE_THROTTLE_TIME_MS);
+        this(curatorFramework, dfs, topic, editLogDir, SimpleConsumerHelper.DEFAULT_REQUIRED_ACKS, DEFAULT_BASE_THROTTLE_TIME_MS);
     }
 
-    public TrumpetLeader(CuratorFramework curatorFramework, DistributedFileSystem dfs, String topic, EditLogDir editLogDir, long baseThrottleTimeMs)
+    public TrumpetLeader(CuratorFramework curatorFramework, DistributedFileSystem dfs, String topic, EditLogDir editLogDir, int kafkaRequiredAcks, long baseThrottleTimeMs)
             throws IOException {
         this.curatorFramework = curatorFramework;
         this.dfs = dfs;
         this.topic = topic;
         this.editLogDir = editLogDir;
+        this.kafkaRequiredAcks = kafkaRequiredAcks;
         this.baseThrottleTimeMs = baseThrottleTimeMs;
     }
 
@@ -82,6 +76,7 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
 
         Metrics.leadershipUptime();
         Metrics.leadershipCounter().inc();
+        Metrics.leadershipStatus().inc();
 
         LOG.debug("Elected as leader, let's stream!");
         boolean forceResetTxId = false;
@@ -107,7 +102,7 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
             File editsLogFile = null;
             long startTxId = 0L;
 
-            producer = getProducer(curatorFramework);
+            producer = getProducer(curatorFramework, kafkaRequiredAcks);
 
             while (run) {
 
@@ -154,6 +149,7 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
                         s = Observable.create(getEditLogObservable(editsLogFile, startTxId))
                                 .subscribe(getProducerSubscriber(topic, producer, lastPublishedTxId));
                         exceptionCounter = 0;
+                        lastTransactionSeenMs = System.currentTimeMillis();
                     } finally {
                         if (s != null) {
                             s.unsubscribe();
@@ -198,11 +194,10 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
                     boolean sameFile = editsLogFile.equals(previousEditLogFile);
                     boolean hasNewTx = startTxId <= lastPublishedTxId.get();
 
-                    if (startTxId > lastPublishedTxId.get() || sameFile) {
+                    if (!hasNewTx || sameFile) {
 
                         if (hasNewTx) {
                             throttleCounter = 1;
-                            lastTransactionSeenMs = System.currentTimeMillis();
                         } else {
                             throttleCounter++;
                         }
@@ -216,7 +211,7 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
                         long sleepMs = Math.min(MAX_THROTTLE_TIME_MS,
                                 baseThrottleTimeMs * (this.random.nextInt(throttleCounter) + 1));
 
-                        LOG.debug("From tx ({}) > to tx ({}), sleeping for {} ms.", startTxId, lastPublishedTxId.get(), sleepMs);
+                        LOG.debug("From tx ({}) > to tx ({}) or sameFile={}, sleeping for {} ms.", startTxId, lastPublishedTxId.get(), sameFile, sleepMs);
 
                         // Measure sleep time
                         Timer.Context sleepTimer = Metrics.sleep().time();
@@ -239,15 +234,20 @@ public class TrumpetLeader extends LeaderSelectorListenerAdapter implements Lead
 
 
         } finally {
+            Metrics.leadershipStatus().dec();
             if (producer != null) {
                 producer.close();
             }
         }
     }
 
-    @VisibleForTesting
-    public Producer getProducer(CuratorFramework curatorFramework) throws Exception {
+    private Producer getProducer(CuratorFramework curatorFramework) throws Exception {
         return SimpleConsumerHelper.getProducer(curatorFramework);
+    }
+
+    @VisibleForTesting
+    public Producer getProducer(CuratorFramework curatorFramework, int requiredAcks) throws Exception {
+        return SimpleConsumerHelper.getProducer(curatorFramework, requiredAcks);
     }
 
     @VisibleForTesting