+++ server/src/main/java/com/verisign/vscc/hdfs/trumpet/server/TrumpetServerCLI.java
@@ -1,19 +1,20 @@
 package com.verisign.vscc.hdfs.trumpet.server;
 
 import com.codahale.metrics.JmxReporter;
-import com.codahale.metrics.MetricFilter;
-import com.codahale.metrics.graphite.Graphite;
-import com.codahale.metrics.graphite.GraphiteReporter;
+import com.codahale.metrics.ScheduledReporter;
 import com.verisign.vscc.hdfs.trumpet.AbstractAppLauncher;
+import com.verisign.vscc.hdfs.trumpet.kafka.SimpleConsumerHelper;
 import com.verisign.vscc.hdfs.trumpet.server.metrics.Metrics;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hdfs.DFSConfigKeys;
 import org.apache.hadoop.util.ToolRunner;
+import org.coursera.metrics.datadog.DatadogReporter;
+import org.coursera.metrics.datadog.transport.UdpTransport;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.InetSocketAddress;
+import java.util.EnumSet;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
@@ -21,18 +22,20 @@ public class TrumpetServerCLI extends AbstractAppLauncher {
 
     public static final String OPTION_DFS_EDITS_DIR = "dfs.edits.dir";
     public static final String OPTION_BASE_THROTTLE_TIME_MS = "base.throttle.time.ms";
-    public static final String OPTION_GRAPHITE_SERVER_HOST_PORT = "graphite.server.hostport";
-    public static final String OPTION_GRAPHITE_PREFIX = "graphite.prefix";
-    public static final String OPTION_GRAPHITE_PREFIX_DEFAULT = "trumpet." + getHostname();
+    public static final String OPTION_KAFKA_REQUIRED_ACKS = "kafka.required.acks";
+    public static final String OPTION_METRICS_PREFIX_DEFAULT = "trumpet." + getHostname();
+    public static final String OPTION_STATSD_REPORTER_ENABLED = "statsd.enable";
+    public static final String OPTION_STATSD_REPORTER_PREFIX = "statsd.prefix";
 
     private final CountDownLatch latch = new CountDownLatch(1);
     private boolean initialized = false;
-    private GraphiteReporter graphiteReporter;
     private TrumpetServer trumpetServer;
     private JmxReporter jmxReporter;
+    private ScheduledReporter metricsReporter;
 
     private File dfsEditsDir;
     private long baseThrottleTimeMs;
+    private int kafkaRequiredAcks;
 
     public static void main(String[] args) throws Exception {
         int res = ToolRunner.run(new Configuration(), new TrumpetServerCLI(), args);
@@ -72,31 +75,45 @@ public class TrumpetServerCLI extends AbstractAppLauncher {
         }
 
         baseThrottleTimeMs = Long.parseLong((String) getOptions().valueOf(OPTION_BASE_THROTTLE_TIME_MS));
+        kafkaRequiredAcks = Integer.parseInt((String) getOptions().valueOf(OPTION_KAFKA_REQUIRED_ACKS));
+
+        boolean statsd = getOptions().has(OPTION_STATSD_REPORTER_ENABLED);
 
         final JmxReporter localJmxReporter = JmxReporter.forRegistry(Metrics.getRegistry()).build();
         jmxReporter = localJmxReporter;
         localJmxReporter.start();
 
-        final GraphiteReporter localGraphiteReporter;
-        if (getOptions().has(OPTION_GRAPHITE_SERVER_HOST_PORT)) {
-            String graphiteServer = (String) getOptions().valueOf(OPTION_GRAPHITE_SERVER_HOST_PORT);
-            String grahitePrefix = (String) getOptions().valueOf(OPTION_GRAPHITE_PREFIX);
-            String[] serverParts = graphiteServer.split(":");
-            final Graphite graphite = new Graphite(
-                    new InetSocketAddress(serverParts[0],
-                            Integer.parseInt(serverParts.length > 1 ? serverParts[1] : "2003")));
-            localGraphiteReporter = GraphiteReporter.forRegistry(Metrics.getRegistry())
-                    .prefixedWith(grahitePrefix)
-                    .convertRatesTo(TimeUnit.SECONDS)
-                    .convertDurationsTo(TimeUnit.MILLISECONDS)
-                    .filter(MetricFilter.ALL)
-                    .build(graphite);
-            localGraphiteReporter.start(1, TimeUnit.MINUTES);
-        } else { localGraphiteReporter = null; }
-        graphiteReporter = localGraphiteReporter;
+        if (statsd) {
+
+            // default prefix (trumpet.$(hostname --fqdn)) was meant for legacy Graphite reporter
+            // if no prefix explicitly set as parameter, using a null prefix instead.
+            String statsdPrefix = null;
+            if (getOptions().has(OPTION_STATSD_REPORTER_PREFIX)) {
+                statsdPrefix = (String) getOptions().valueOf(OPTION_STATSD_REPORTER_PREFIX);
+            }
+
+            EnumSet<DatadogReporter.Expansion> expansions = EnumSet.of(DatadogReporter.Expansion.COUNT,
+                    DatadogReporter.Expansion.RATE_1_MINUTE,
+                    DatadogReporter.Expansion.RATE_15_MINUTE,
+                    DatadogReporter.Expansion.MEDIAN,
+                    DatadogReporter.Expansion.P95,
+                    DatadogReporter.Expansion.P99);
+
+            UdpTransport udpTransport = new UdpTransport.Builder().build();
+            metricsReporter = DatadogReporter.forRegistry(Metrics.getRegistry())
+//                    .withEC2Host()
+                    .withTransport(udpTransport)
+//                    .withTags()
+                    .withExpansions(expansions)
+                    .withPrefix(statsdPrefix)
+                    .build();
+
+            metricsReporter.start(10, TimeUnit.SECONDS);
+        }
+
         Metrics.uptime();
 
-        final TrumpetServer localTrumpetServer = new TrumpetServer(getCuratorFrameworkUser(), getConf(), getTopic(), dfsEditsDir, baseThrottleTimeMs);
+        final TrumpetServer localTrumpetServer = new TrumpetServer(getCuratorFrameworkUser(), getConf(), getTopic(), dfsEditsDir, kafkaRequiredAcks, baseThrottleTimeMs);
         trumpetServer = localTrumpetServer;
         Runtime.getRuntime().addShutdownHook(new Thread() {
 
@@ -106,20 +123,13 @@ public class TrumpetServerCLI extends AbstractAppLauncher {
                 System.out.println("Trumpet shutdown from external signal received.");
 
                 try {
-                    localTrumpetServer.close();
+                    internalClose();
                 } catch (IOException e) {
-                    // well, can't really recover here, shutting down anyway...
+                    System.out.println("Exception while closing everythin. Cannot recover anyway, completing the shutdown process...");
                     e.printStackTrace();
-                } finally {
-                    localJmxReporter.stop();
-                    if (localGraphiteReporter != null) {
-                        localGraphiteReporter.stop();
-                    }
-                    Metrics.close();
                 }
             }
         });
-
 
         initialized = true;
 
@@ -142,10 +152,11 @@ public class TrumpetServerCLI extends AbstractAppLauncher {
         getParser().accepts(OPTION_BASE_THROTTLE_TIME_MS, "Min throttle time increment when no transactions are found. " +
                 "The bigger value, the less resources will be used but the more time before events might be published.")
                 .withRequiredArg().defaultsTo(String.valueOf(TrumpetServer.DEFAULT_BASE_THROTTLE_TIME_MS));
-        getParser().accepts(OPTION_GRAPHITE_SERVER_HOST_PORT, "Graphite server host and port where to report metrics, if set. Port default to 2003 if not set.")
-                .withRequiredArg();
-        getParser().accepts(OPTION_GRAPHITE_PREFIX, "Prefix to add on the metrics reported to Graphite.")
-                .withRequiredArg().defaultsTo(OPTION_GRAPHITE_PREFIX_DEFAULT);
+        getParser().accepts(OPTION_STATSD_REPORTER_ENABLED, "Enable statsd report to the local agent (port 8215)");
+        getParser().accepts(OPTION_STATSD_REPORTER_PREFIX, "Prefix to add on the metrics reported to statsd, if enabled.")
+                .withRequiredArg().defaultsTo(OPTION_METRICS_PREFIX_DEFAULT);
+        getParser().accepts(OPTION_KAFKA_REQUIRED_ACKS, "Kafka request.required.acks property. 0=none (async), 1=leader, -1=all. See kafka documentation for more details.")
+                .withRequiredArg().defaultsTo(String.valueOf(SimpleConsumerHelper.DEFAULT_REQUIRED_ACKS));
     }
 
 
@@ -161,8 +172,8 @@ public class TrumpetServerCLI extends AbstractAppLauncher {
             e.printStackTrace();
         } finally {
             jmxReporter.stop();
-            if (graphiteReporter != null) {
-                graphiteReporter.stop();
+            if (metricsReporter != null) {
+                metricsReporter.stop();
             }
             Metrics.close();
         }