+++ server/src/main/java/com/verisign/vscc/hdfs/trumpet/server/editlog/EditLogDir.java
@@ -41,9 +41,6 @@ public class EditLogDir {
             @Override
             public boolean accept(File pathname) {
                 if (pathname.isDirectory()) {
-                    if (CURRENT.equals(pathname.getName())) {
-                        return true;
-                    }
                     File[] currents = pathname.listFiles(new FilenameFilter() {
                         @Override
                         public boolean accept(File dir, String name) {
@@ -61,11 +58,8 @@ public class EditLogDir {
                 " " + rootDir + "/*/current.");
         Preconditions.checkState(editLogDirs.length == 1, "No folder or too many folders matching edit log dir structure in rootDir:" +
                 " " + rootDir + "/*/current.");
-        if (CURRENT.equals(editLogDirs[0].getName())) {
-            this.editLogDir = editLogDirs[0];
-        } else {
-            this.editLogDir = new File(editLogDirs[0], CURRENT);
-        }
+
+        this.editLogDir = new File(editLogDirs[0], CURRENT);
 
         Preconditions.checkState(this.editLogDir.isDirectory());
         Preconditions.checkState(this.editLogDir.canRead());